Some repairs

+++ app/helpers/application_helper.rb
@@ -1,12 +1,24 @@
 module ApplicationHelper
 
   def edit_and_destroy_buttons(item)
-    unless current_user.nil?
+  unless current_user.nil?
+    if current_user.admin?
       edit = link_to('Edit', url_for([:edit, item]), class:"btn btn-primary")
       del = link_to('Destroy', item, method: :delete,
                     data: {confirm: 'Are you sure?' }, class:"btn btn-danger")
       raw("#{edit} #{del}")
     end
+   end
+  end
+
+  def new_buttons(item)
+  unless current_user.nil?
+    if current_user.admin?
+      byebug
+      new = link_to('New', url_for([:new, item]), class:"btn btn-default")
+      raw("#{new}")
+    end
+   end
   end
 
  def favorite_button(post)
